Create Array Reversal Program.java
This Java program reverses the elements of a one-dimensional integer array. It swaps the first and last elements, the second and second-to-last elements, and so on, until the entire array is reversed. The program uses a for loop with two pointers (i and j) that move inward from both ends of the array. The reversed array is then displayed using Arrays.toString().