Create Sorting an Array Using Quick Sort.c
Sorting an Array Using Quick Sort

This C program demonstrates the use of the qsort function from the standard library to sort an array of integers in ascending order. It defines a custom comparator function to compare two integers and utilizes qsort to rearrange the elements. The sorted array is then printed to the console.