Create Coin Change.c
Coin Change - Greedy Algorithm Approach

This program is based on greedy approach. The goal is to find the minimum number of coins needed to make a given amount a using denominations provided in the array coins. The array is traversed from the largest denomination to the smallest, deducting the coin's value from the amount until the amount becomes zero. The program outputs the coin denominations used to make up the total.